+++ cmd/restic/integration_helpers_test.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"syscall"
 	"testing"
 
 	. "github.com/restic/restic/test"
@@ -68,38 +67,6 @@ func sameModTime(fi1, fi2 os.FileInfo) bool {
 	}
 
 	return fi1.ModTime() == fi2.ModTime()
-}
-
-func (e *dirEntry) equals(other *dirEntry) bool {
-	if e.path != other.path {
-		fmt.Fprintf(os.Stderr, "%v: path does not match (%v != %v)\n", e.path, e.path, other.path)
-		return false
-	}
-
-	if e.fi.Mode() != other.fi.Mode() {
-		fmt.Fprintf(os.Stderr, "%v: mode does not match (%v != %v)\n", e.path, e.fi.Mode(), other.fi.Mode())
-		return false
-	}
-
-	if !sameModTime(e.fi, other.fi) {
-		fmt.Fprintf(os.Stderr, "%v: ModTime does not match (%v != %v)\n", e.path, e.fi.ModTime(), other.fi.ModTime())
-		return false
-	}
-
-	stat, _ := e.fi.Sys().(*syscall.Stat_t)
-	stat2, _ := other.fi.Sys().(*syscall.Stat_t)
-
-	if stat.Uid != stat2.Uid {
-		fmt.Fprintf(os.Stderr, "%v: UID does not match (%v != %v)\n", e.path, stat.Uid, stat2.Uid)
-		return false
-	}
-
-	if stat.Gid != stat2.Gid {
-		fmt.Fprintf(os.Stderr, "%v: GID does not match (%v != %v)\n", e.path, stat.Gid, stat2.Gid)
-		return false
-	}
-
-	return true
 }
 
 // directoriesEqualContents checks if both directories contain exactly the same
@@ -237,6 +204,8 @@ func withTestEnvironment(t testing.TB, f func(*testEnvironment, GlobalOptions)) 
 	}
 
 	OK(t, os.MkdirAll(env.testdata, 0700))
+	OK(t, os.MkdirAll(env.cache, 0700))
+	OK(t, os.MkdirAll(env.repo, 0700))
 
 	f(&env, configureRestic(t, env.cache, env.repo))
